fix(ios,android): ProgressBar max/value normalization, split AdvancedElements into 1:1 files
- Add `max` property to ProgressBar model (iOS + Android) for absolute value scales
- Add `normalizedValue` computed property: auto-detects 0–1 fractional vs 0–100 scale
- Update renderers to use normalizedValue for bar width, percentage text, accessibility
- Split Android AdvancedElements.kt monolith into 9 focused files matching iOS structure:
  Accordion, Carousel, ChartElements, CodeBlock, CompoundButton, ListElement,
  ProgressElements, RatingElements, TabSet
