+++ android/ac-core/src/main/kotlin/com/microsoft/adaptivecards/core/models/ListElement.kt
@@ -1,0 +1,22 @@
+package com.microsoft.adaptivecards.core.models
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+import kotlinx.serialization.Transient
+import kotlinx.serialization.json.JsonElement
+
+@Serializable
+@SerialName("List")
+data class ListElement(
+    @Transient override val type: String = "List",
+    override val id: String? = null,
+    override val isVisible: Boolean = true,
+    override val separator: Boolean = false,
+    override val spacing: Spacing? = null,
+    override val height: BlockElementHeight? = null,
+    override val requires: Map<String, String>? = null,
+    override val fallback: JsonElement? = null,
+    val items: List<CardElement>,
+    val maxHeight: String? = null,
+    val style: String? = null
+) : CardElement
